drl-assembly-planning-main/requirements.txt
absl-py
cachetools
charset-normalizer
click
cloudpickle
docker-pycreds
gitdb
GitPython
google-auth
google-auth-oauthlib
grpcio
gym==0.25.2
gym-notices
h5py
idna
importlib-metadata
llvmlite>=0.40.0
Markdown
MarkupSafe
numba>=0.57.0
numpy>=1.23.5
oauthlib
pathtools
promise
protobuf
psutil
pyasn1
pyasn1-modules
pygame
PyOpenGL
PyYAML
requests
requests-oauthlib
rsa
sentry-sdk
setproctitle
shortuuid
six
smmap
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
tianshou==0.4.9.post1
torch>=2.0.0
tqdm
typing_extensions
urllib3
wandb
Werkzeug
zipp
pettingzoo
